+++ packages/http-server/src/getHttpConfigFromRequest.ts
@@ -6,11 +6,16 @@ import * as D from 'io-ts/lib/Decoder';
 import * as parsePreferHeader from 'parse-prefer-header';
 
 const BooleanFromString = D.parse<string, boolean>(s =>
-  s === 'true' ? D.success(true) : s === 'false' ? D.success(false) : D.failure(s, 'Not a boolean')
+  s === 'true' ? D.success(true) : s === 'false' ? D.success(false) : D.failure(s, 'a boolean')
 );
 
+const IntegerFromString = D.parse<string, number>(s => {
+  const parsedString = parseInt(s, 10);
+  return !isNaN(parsedString) ? D.success(parsedString) : D.failure(s, 'a number');
+});
+
 const PreferencesDecoder = D.partial({
-  code: D.string,
+  code: pipe(D.string, IntegerFromString),
   dynamic: pipe(D.string, BooleanFromString),
   example: D.string,
 });